Id typ wieder von long auf int zurückentwicklelt

+++ src/main/java/at/ac/tuwien/sepm/ws16/qse01/dao/LogoDAO.java
@@ -37,7 +37,7 @@ public interface LogoDAO {
      * exist in persistence store, null value will be returned
      * @throws PersistenceException if persistence data store can not be accessed
      */
-    Logo read(long id) throws PersistenceException;
+    Logo read(int id) throws PersistenceException;
 
     /**
      * Retrieve all logo objects from persistence store(that have not been deleted)
